+++ CHANGELOG.md
@@ -5,11 +5,16 @@
 ### Added
 
 - API Reference now can be found [here](https://cdn.bitmovin.com/player/reactnative/0/docs/index.html)
+- `PictureInPictureConfig` to `PlayerView` to allow configuring the Picture in Picture behavior
 
 ### Changed
 
 - Update IMA SDK dependency on Android to `3.31.0`
 - Update Bitmovin's native Android SDK version to `3.47.0`
+
+### Deprecated
+
+- `PlaybackConfig.isPictureInPictureEnabled` in favor of `PictureInPictureConfig.isEnabled`
 
 
 ## [0.12.0] (2023-09-25)